requirements.txt
pyodbc
psycopg2-binary
python-dotenv